moved helper functions into separate module for #35
